Use animation and appearance configuration for MessageView
